MVC Implemented
